test(iroh-net) add simple roundtrip test for frame encoding (#1397)
## Description

Add a simple roundtrip test for the frame encoding

## Notes & open questions

This adds just a happy case test. I was going to also add failure tests,
but I noticed that I don't really understand the protocol well enough.
E.g. which variable size content must be below MAX_PACKET_SIZE.

E.g. for the message types that contain an encrypted message: is there
some size limit on the encrypted_message, and what exactly is it?

For RecvPacket, what is the size limit on the content? Is it
MAX_PACKET_SIZE or MAX_PACKET_SIZE - PUBLIC_KEY_LENGTH?

## Change checklist

- [x] Self-review.
- [x] Tests if relevant.