chore(ci): Configure sccache to use local caches (#1865)
## Description

Because we use self-hosted runners we do not want the sccache action
to query the github action cache service for each compilation request.
Instead it should be using the local cache.

This means the first time something is compiled on a runner it will
not yet be in the cache, even if another runner already did compile
this.  However because the cache size is set reasonably large over
time all the common dependencies will end up cached on each
self-hosted runner.

Maybe it will be possible to run a self-hosted caching service
sometime, but for now relying on a large local cache is expected to be
reasonable.

## Notes & open questions

- This will not start showing results right away, as caches on the
  self-hosted runners need to be filled.  They are full by now though
  and show improvements across the board.
  
- Please check that all local runners have 100G space for a cache in
  `~/.cache/sccache` or `%LOCALAPPDATA%\Mozilla\sccache`.

## Change checklist

- [x] Self-review.
- ~~Documentation updates if relevant.~~
- ~~Tests if relevant.~~