chore(ci): Set up a separate workflow for flaky tests (#1922)
## Description

This sets up a separate workflow which runs all tests, including flaky
ones. It can be triggered by setting the "flaky-test" label on a PR,
which will disable most of the normal CI workflow and instead run just
the tests including flaky ones.

The flaky workflow can also be manually triggered on a specific branch.

## Notes & open questions

- The flaky workflow only runs the tests, none of the other jobs.
- Cross-compiled jobs aren't even included. Maybe they should be but
right now they don't run tests at all so is pointless.
- On any labelling change for the PR the jobs in the workflows will be
cancelled if they're still running. I don't think we generally label PRs
so that's probably fine.
- Flaky tests are going to be carnage for a while!
- We have no notification for the cron job results yet. Not sure how to
approach this. I'd be fine doing this later.
- The `workflow_dispatch` version (aka manual trigger) is untested, I
think we have to wait until it is merged to main to be able to test
that?
- The UI will now always show all jobs: the skipped flaky jobs on normal
PRs and the missing normal jobs on flaky PRs. That's probably fine? A
bit noisier though.
- The set of "required jobs" has changed again and will need patching up
in branch protection.

This implements the bulk of
https://github.com/n0-computer/iroh/issues/1919

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- [x] Tests if relevant.

---------

Co-authored-by: Asmir Avdicevic <asmir.avdicevic64@gmail.com>