fix(iroh-net): Fix some flaky magicsock tests (#2034)
## Description

Fix flaky tests in iroh-net (hopefully).  The main issue was that the
testing code meshing together the sockets had a race-condition and
would sometimes fail.

Other improvements:

- Logging improvements of various parts.  These tests should now
  provide better info on failure.

- Add alternative formatting of NodeId to use the short format.

- Use NodeId in the public API, rather than PublicKey.

- Move multi-threaded logging setup to the testing utils crate now
  that it can work properly.  More code will have to switch.

## Notes & open questions

The local_endpoints_change API is really difficult to use correctly.
Even the way this testing code uses it now is wrong.  Instead of the
two current APIs there should be a single API which returns a stream.
I'll do this as a separate followup PR though.

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- [x] Tests if relevant.