refactor: move `iroh_net` base types to `iroh_base` (#2053)
## Description

This moves the basic types from iroh_net to iroh_base, so that crates do
not have to depend on iroh_net if they just need to interface with its
basic types.

It moves the following:
* `iroh_net::key` module with the `NodeId`/`PublicKey` and `SecretKey`
types, plus the crypto dependencies for those types. Because this adds
weight to iroh_base they are behind a non-default `key` feature there.
* `iroh_net::ticket::{BlobTicket, NodeTicket}` - they were only in
iroh_net because we couldn't move them because of their dependendy on
`NodeId`
* `iroh_net::AddrInfo` and `NodeAddr` - are used by the tickets, and
also very simple types

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [ ] Self-review.
- [ ] Documentation updates if relevant.
- [ ] Tests if relevant.