refactor(iroh)!: replace public fields in iroh client with accessors and use ref-cast to eliminate them entirely (#2350)
## Description

With 4 different clients, the current approach might be OK. But we are
going to have more. E.g. gossip, see
https://github.com/n0-computer/iroh/pull/2258 .

And in any case it feels weird to store the same thing multiple times.

So this replaces public fields in the iroh client with accessors and use
ref-cast to eliminate them entirely. There is now only one rpc field,
and you can switch from that to the different subsystem client views
without runtime overhead, not even an arc clone.

## Breaking Changes

Everything that uses the iroh client will have to switch from field
accesses to accessor fns.

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [ ] Self-review.
- [ ] Documentation updates if relevant.
- [ ] Tests if relevant.
- [ ] All breaking changes documented.