refactor(iroh-net)!: Improve server modules structure & rename structs (#2568)
## Description

<!-- A summary of what this pull request achieves and a rough list of
changes. -->
This is moving around the relay modules a bit. Mostly so that all
server-specific code is under `iroh-net/src/relay/server/`.
This means there's only a handful of `#[cfg(feature = "iroh-relay")]`s
in `relay.rs` and a bunch of other places due to unused warnings
otherwise.

I've also renamed what was previously `iroh::relay::client::{Client,
ClientBuilder, ClientReceiver, ...}` to
`iroh::relay::client::conn::{Conn, ConnBuilder, ConnReceiver, ...}`,
because otherwise the `Client*` names were duplicated in two modules
that were in a parent-child relationship, which is super confusing.
And really, the inner `Client`s are more about individual connections.

I also resolved a TODO around moving to an `AbortingJoinHandle` in favor
of a custom `RelayHttpServerGuard`.

## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->
- Refactored the module structure for users of the `iroh-relay` feature
in iroh-net.

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->
I realize the diff here is quite substantial. Unfortunately the diff is
mostly just bad because I ended up "swapping places" for modules in two
cases, which ended up creating really bad diffs.

I know the diff is quite sad, so it's fair if we say "this is
unreviewable" and I scrap this piece of work. I'm not blocked on this
merging, and I only fairly quickly knocked this out today.
As Mark Twain famously said
> I apologize for such a long PR - I didn't have time to write a short
one.

Or sth like that.

## Change checklist

- [X] Self-review.
- [X] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- ~~[ ] Tests if relevant.~~
- [X] All breaking changes documented. Yeah? Kinda?