chore(ci): use nextests groups to isolate some tests (#2617)
## Description

After a lot of debugging we realized the integrations tests (`gc` tests,
for
example) run with swarm discovery on since they are not technically
using test
iroh, so they bypass the cfg that prevents it from being turned on. This
causes
the `test_local_swarm_discovery` test to be flooded with nodes from
other
tests, making it flaky. Isolating the test should make it no longer
flaky
anymore

## Breaking Changes

n/a

## Notes & open questions

We should still create a test that creates a bunch of nodes and makes
sure they
all can find each other within reasonable time. There is still reason to
believe swarm discovery is slow, particularly in linux, but we should
address
it in a test under our countrol, not via the non-deterministic nature of
concurrent tests in ci.

## Change checklist

- [x] Self-review.
- [ ] ~~Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.~~
- [ ] ~~Tests if relevant.~~
- [ ] ~~All breaking changes documented.~~