docs(iroh): Enable iroh_docsrs feature (#2780)
## Description

This uses the same iroh_docsrs cfg as the other crates and updates the
preview CI job to also use this.

## Breaking Changes

None

## Notes & open questions

However, the world is a sad place.

As far as I can figure out this all correctly annotates the features in
our docs in the individual crates. That is, `https://docs.rs/iroh-net`
and `https://docs.rs/iroh-gossip` etc. This is also true for the `iroh`
crate. BUT, the iroh crate uses `doc(inline)` on the other crates, and
when using this we can not get the feature annotation in the sub-crates.
Sad times.

## Change checklist

- [x] Self-review.
- ~~[ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.~~
- ~~[ ] Tests if relevant.~~
- ~~[ ] All breaking changes documented.~~