feat: Added option to choose a specific socket_addr for connect and accept (#43)
## Description

This is a small optional addition to the iroh-doctor accept / connect
commands. It adds a new `socket_addr` flag that is used in the Endpoint
configurations `bind_addr_(v4/v6)` methods. With this it is possible to
have a more fine grained control which socket_address is used when
testing connections.

I was in need of this feature while investigating: 
https://github.com/n0-computer/iroh/discussions/3380

## Breaking Changes

None, the flag is completely optional

## Notes & open questions

None

## Change checklist

- [x] Self-review.
- [ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.
- [ ] All breaking changes documented.